+++ appObjects/FlatCAMExcellon.py
@@ -1323,8 +1323,10 @@ class ExcellonObject(FlatCAMObj, Excellon):
                             t = f_tool
                             break
                     if t:
-                        fused_tools_dict[t]['drills'] += tool_dict['drills']
-                        fused_tools_dict[t]['slots'] += tool_dict['slots']
+                        if 'drills' in tool_dict and tool_dict['drills']:
+                            fused_tools_dict[t]['drills'] += tool_dict['drills']
+                        if 'slots' in tool_dict and tool_dict['slots']:
+                            fused_tools_dict[t]['slots'] += tool_dict['slots']
                         fused_tools_dict[t]['solid_geometry'] += tool_dict['solid_geometry']
                 else:
                     fused_tools_dict[toolid] = tool_dict